# src/auth/dto/signin-dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, IsEmail, MinLength } from 'class-validator';

export class SignInDto {
  @ApiProperty()
  @IsEmail()
  email: string;

  @IsString()
  @MinLength(5)
  @ApiProperty()
  password: string;
}