# swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/signin":{"post":{"operationId":"AuthController_signIn","summary":"Логин","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}}}}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","summary":"Регистрация","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}}}}},"/user":{"get":{"operationId":"UserController_getProfile","summary":"Получение текущего пользователя","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_delete","summary":"Удаление аккаунта","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteResult"}}}}},"security":[{"bearer":[]}]}},"/todo":{"post":{"operationId":"TodoController_create","summary":"Создание todo","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}},"security":[{"bearer":[]}]}},"/todo/{id}":{"put":{"operationId":"TodoController_update","summary":"Изменение todo по id","parameters":[{"name":"id","required":true,"in":"path","description":"Идентификатор todo","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}},"security":[{"bearer":[]}]},"delete":{"operationId":"TodoController_delete","summary":"Удаление todo по id","parameters":[{"name":"id","required":true,"in":"path","description":"Идентификатор todo","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteResult"}}}}},"security":[{"bearer":[]}]}},"/todo/list/{status}":{"get":{"operationId":"TodoController_findAllByStatus","summary":"Получение списка todo по статусу","parameters":[{"name":"status","required":true,"in":"path","description":"Статус todo","schema":{"type":"string","enum":["CREATED","IN_PROGRESS","COMPLETED"]}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Todo"}}}}}},"security":[{"bearer":[]}]}}},"info":{"title":"Todo list API","description":"Todo list API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"User":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"}},"required":["id","email"]},"DeleteResult":{"type":"object","properties":{}},"CreateTodoDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title"]},"TodoStatus":{"type":"string","enum":["CREATED","IN_PROGRESS","COMPLETED"]},"Todo":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"status":{"$ref":"#/components/schemas/TodoStatus"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdBy":{"$ref":"#/components/schemas/User"}},"required":["id","title","status","createdAt","updatedAt","createdBy"]},"UpdateTodoDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"$ref":"#/components/schemas/TodoStatus"}}}}}}